myapp/requirements.txt


alembic==1.13.1
async-timeout==4.0.3
backports-abc==0.5
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
dominate==2.9.1

flask==3.0.3

Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
flask-redis==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
livereload==2.6.3
Mako==1.3.2
MarkupSafe==2.1.5
meld3==2.0.1
mysqlclient==2.2.3
packaging==23.2
pycparser==2.21
pycrypto==2.6.1
PyMySQL==1.1.0
python-dotenv==1.0.1
python-editor==1.0.4
redis==5.0.1
requests==2.31.0
shellescape==3.8.1
simplejson==3.19.2
singledispatch==4.1.0
six==1.16.0
SQLAlchemy==2.0.25
supervisor==4.2.5
tornado==6.4
typing_extensions==4.9.0
urllib3==2.2.0
uWSGI==2.0.23
visitor==0.1.3
Werkzeug==3.0.1
WTForms==3.1.2

# OpenTelemetry Core and Instrumentation
prometheus-client==0.20.0
opentelemetry-distro==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-logging==0.46b0
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-sqlalchemy==0.46b0
flask_prometheus_metrics[flask]
opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-requests

